pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tradingbot"
version = "0.1.0"
description = "Multi-exchange trading bot for crypto, stocks, and prediction markets"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.2.0",
    "py-clob-client>=0.20.0",
    "alpaca-py>=0.21.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "tomli>=2.0.1",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
tradingbot = "tradingbot.cli.commands:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/tradingbot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
